backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic-settings>=2.0
pydantic[email]>=2.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-docx>=1.1
PyMuPDF>=1.24
qrcode[pil]>=8.0
pyHanko>=0.25,<0.26
python-multipart>=0.0.9
